+++ CHANGELOG.md
@@ -39,7 +39,8 @@ to communicate reliably.
 - Agent-level hyper-parameter to enable quantisation for self-hosted models.
 - Using Liger kernel in DPO training for increased speed and lower memory usage.
 - Now logging prompt and completion token lengths for self-hosted models.
-- Now attempting several times to start the vLLM server. 
+- Now attempting several times to start the vLLM server.
+- Allowed setting DPO number of epochs.
 
 
 ### Changed
@@ -52,6 +53,7 @@ to communicate reliably.
   run`), which has allowed displaying more log messages.
 - Sanitising model name a little more when creating training job name, which helps
   Hugging Face identify the base model more easily.
+- Allowed setting DPO training logging period, and set it to 1 by default.
 
 
 ### Fixed